refactor(ecology): keep wet placement logic in strategies
